+++ packages/create-svelte/test/check.js
@@ -1,10 +1,10 @@
-import fs from 'fs';
-import { execSync } from 'child_process';
-import path from 'path';
+import fs from 'node:fs';
+import { execSync } from 'node:child_process';
+import path from 'node:path';
 import { test } from 'uvu';
 import * as assert from 'uvu/assert';
 import { create } from '../index.js';
-import { fileURLToPath } from 'url';
+import { fileURLToPath } from 'node:url';
 import glob from 'tiny-glob';
 // use a directory outside of packages to ensure it isn't added to the pnpm workspace
 const test_workspace_dir = fileURLToPath(
@@ -21,7 +21,10 @@ const overrides = { ...existing_workspace_overrides };
 	await glob(fileURLToPath(new URL('../../../packages', import.meta.url)) + '/*/package.json')
 ).forEach((pkgPath) => {
 	const name = JSON.parse(fs.readFileSync(pkgPath, 'utf-8')).name;
-	overrides[name] = path.dirname(path.resolve(pkgPath));
+	// use `file:` protocol for opting into stricter resolve logic which catches more bugs,
+	// but only on CI because it doesn't work locally for some reason
+	const protocol = process.env.CI ? 'file:' : '';
+	overrides[name] = `${protocol}${path.dirname(path.resolve(pkgPath))}`;
 });
 
 try {