+++ packages/package/src/typescript.js
@@ -1,8 +1,8 @@
-import * as fs from 'fs';
-import * as path from 'path';
-import { createRequire } from 'module';
+import * as fs from 'node:fs';
+import * as path from 'node:path';
+import { createRequire } from 'node:module';
 import { posixify, mkdirp, rimraf, walk } from './filesystem.js';
-import { resolve_lib_alias, write } from './utils.js';
+import { resolve_aliases, write } from './utils.js';
 import { emitDts } from 'svelte2tsx';
 
 /**
@@ -10,35 +10,30 @@ import { emitDts } from 'svelte2tsx';
  * The files are written to a temporary location and those which should be kept
  * are sanitized ($lib alias resolved) and copied over to the destination folder.
  *
- * @param {import('./types').ValidatedConfig} config
+ * @param {string} input
+ * @param {string} output
  * @param {string} cwd
+ * @param {Record<string, string>} alias
  * @param {import('./types').File[]} files
  */
-export async function emit_dts(config, cwd, files) {
-	const tmp = `${config.package.dir}/__package_types_tmp__`;
+export async function emit_dts(input, output, cwd, alias, files) {
+	const tmp = `${output}/__package_types_tmp__`;
 	rimraf(tmp);
 	mkdirp(tmp);
 
 	const require = createRequire(import.meta.url);
 	await emitDts({
-		libRoot: config.package.source,
+		libRoot: input,
 		svelteShimsPath: require.resolve('svelte2tsx/svelte-shims.d.ts'),
 		declarationDir: path.relative(cwd, tmp)
 	});
 
 	const handwritten = new Set();
-	const excluded = new Set();
 
-	// remove excluded files, and files that conflict with hand-written .d.ts
+	// skip files that conflict with hand-written .d.ts
 	for (const file of files) {
 		if (file.name.endsWith('.d.ts')) {
 			handwritten.add(file.name);
-		}
-
-		if (!file.is_included) {
-			excluded.add(file.base + '.d.ts');
-			excluded.add(file.base + '.d.mts');
-			excluded.add(file.base + '.d.cts');
 		}
 	}
 
@@ -50,11 +45,8 @@ export async function emit_dts(config, cwd, files) {
 			console.warn(`Using $lib/${normalized} instead of generated .d.ts file`);
 		}
 
-		// don't overwrite hand-written .d.ts files
-		if (excluded.has(normalized)) continue;
-
 		const source = fs.readFileSync(path.join(tmp, normalized), 'utf8');
-		write(path.join(config.package.dir, normalized), resolve_lib_alias(normalized, source, config));
+		write(path.join(output, normalized), resolve_aliases(input, normalized, source, alias));
 	}
 
 	rimraf(tmp);