+++ packages/package/test/fixtures/exports-merge/expected/internal/_Private.svelte
@@ -1,1 +1,0 @@
-This is a private component and should be excluded